refactor: overhaul docs-extractor architecture for modular type extraction
- Removed legacy order.json configuration file.
- Enhanced types.ts with new interfaces for component and prop information.
- Updated tsconfig.json to include path mapping for improved module resolution.
- Created comprehensive README.md detailing the new architecture and usage.
- Implemented CLI tool for extracting component types from specified files.
- Developed ComponentAnalyzer for extracting display names and default elements from React components.
- Introduced ExternalTypeResolver for resolving external type definitions from npm packages.
- Built PropsAnalyzer for extracting prop information, including types and default values.
- Refactored TypeExtractor to orchestrate the extraction process using various analyzers.
- Added utility functions for TypeScript analysis, including JSDoc extraction and type parsing.
- Integrated Vanilla Extract support for CSS-in-JS styling analysis.
