--- packages/docs-extractor/src/cli-standalone.ts
@@ -1,7 +1,0 @@
-#!/usr/bin/env node
-
-import { createCliCommand } from './cli';
-import { main } from './main';
-
-// CLI 실행
-createCliCommand(main).parse();
