+++ packages/docs-extractor/src/component.ts
