+++ packages/docs-extractor/src/index.ts
@@ -1,18 +1,18 @@
 // Type Extraction Core
-export { TypeExtractor, extractComponentTypesFromFile } from './lib/type-extractor';
+export { TypeExtractor, extractComponentTypesFromFile } from '~/type';
 
 // Specialized Analyzers
 
-export { ComponentAnalyzer } from './lib/component-analyzer';
-export { PropsAnalyzer } from './lib/props-analyzer';
-export { VanillaExtractAnalyzer } from './lib/vanilla-extract-analyzer';
+export { ComponentAnalyzer } from '~/component';
+export { PropsAnalyzer } from '~/props';
+export { VanillaExtractAnalyzer } from '~/vanilla-extract';
 
 // External Type Resolution
-export { ExternalTypeResolver } from './lib/external-type-resolver';
-export type { ExternalPackageConfig } from './lib/external-type-resolver';
+export { ExternalTypeResolver } from '~/external';
+export type { ExternalPackageConfig } from '~/external';
 
 // Types and Interfaces
-export type { ComponentTypeInfo, PropInfo, TypeExtractorConfig } from './lib/types';
+export type { ComponentTypeInfo, PropInfo, TypeExtractorConfig } from '~/types';
 
 // Utilities
 export {
@@ -23,7 +23,4 @@ export {
     getJSDocDefaultValue,
     getLiteralValue,
     parseTypeToArray,
-} from './lib/utils';
-
-// Main function for convenience
-export { main } from './main';
+} from '~/utils';
