--- packages/docs-extractor/src/order.json
@@ -1,104 +1,0 @@
-{
-    "cssVariables": [],
-    "dataAttributes": [
-        "data-checked",
-        "data-unchecked",
-        "data-open",
-        "data-closed",
-        "data-panel-open",
-        "data-popup-open",
-        "data-pressed",
-        "data-selected",
-        "data-highlighted",
-        "data-dragging",
-        "data-orientation",
-        "data-disabled",
-        "data-readonly",
-        "data-required",
-        "data-valid",
-        "data-invalid",
-        "data-dirty",
-        "data-touched",
-        "data-uncentered",
-        "data-anchor-hidden",
-        "__EVERYTHING_ELSE__",
-        "data-starting-style",
-        "data-ending-style"
-    ],
-    "props": [
-        "aria-label",
-        "aria-labelledby",
-        "name",
-        "label",
-        "defaultChecked",
-        "checked",
-        "onCheckedChange",
-        "indeterminate",
-        "defaultValue",
-        "value",
-        "onValueChange",
-        "onValueCommitted",
-        "allValues",
-        "defaultOpen",
-        "open",
-        "onOpenChange",
-        "defaultPressed",
-        "pressed",
-        "onPressedChange",
-        "onPressedCommitted",
-        "onClick",
-        "closeOnClick",
-        "errors",
-        "onClearErrors",
-        "hidden",
-        "hiddenUntilFound",
-        "__EVERYTHING_ELSE__",
-        "step",
-        "smallStep",
-        "largeStep",
-        "minStepsBetweenValues",
-        "min",
-        "max",
-        "allowWheelScrub",
-        "format",
-        "autoFocus",
-        "container",
-        "align",
-        "alignOffset",
-        "side",
-        "sideOffset",
-        "arrowPadding",
-        "anchor",
-        "collisionBoundary",
-        "collisionPadding",
-        "sticky",
-        "positionMethod",
-        "trackAnchor",
-        "trackCursorAxis",
-        "initialFocus",
-        "finalFocus",
-        "match",
-        "forceShow",
-        "disabled",
-        "readOnly",
-        "required",
-        "valid",
-        "invalid",
-        "openOnHover",
-        "delay",
-        "closeDelay",
-        "timeout",
-        "hoverable",
-        "loop",
-        "orientation",
-        "inputRef",
-        "validate",
-        "validationMode",
-        "validationDebounceTime",
-        "id",
-        "children",
-        "className",
-        "keepMounted",
-        "render"
-    ]
-}
