+++ packages/docs-extractor/src/types.ts
@@ -1,3 +1,6 @@
+/**
+ * Common types and interfaces for the type extraction system
+ */
 import ts from 'typescript';
 import * as tae from 'typescript-api-extractor';
 
@@ -6,6 +9,7 @@ export interface RunOptions {
     configPath: string;
     out: string;
     language: string;
+    externalTypePaths?: string[];
 }
 
 export interface TsConfig {
@@ -17,4 +21,32 @@ export interface ExportProcessingResult {
     exports: tae.ExportNode[];
     errorCount: number;
     fileExportsMap: Map<string, tae.ExportNode[]>;
+}
+
+export interface ComponentTypeInfo {
+    name: string;
+    displayName?: string;
+    description?: string;
+    props: PropInfo[];
+    defaultElement?: string;
+}
+
+export interface PropInfo {
+    name: string;
+    type: string | string[];
+    required: boolean;
+    description?: string;
+    defaultValue?: string;
+}
+
+export interface BaseUIComponent {
+    modulePath: string;
+    componentName: string;
+}
+
+export interface TypeExtractorConfig {
+    configPath: string;
+    files?: string[];
+    projectRoot?: string;
+    externalTypePaths?: string[];
 }
